+++ GameEngine/src/math/vec4.h
@@ -25,47 +25,47 @@ namespace cgl
 		{
 			struct
 			{
-				double x;
-				double y;
-				double z;
-				double w;
+				float x;
+				float y;
+				float z;
+				float w;
 			};
-			std::array<double, 4> e{0, 0, 0, 0};
+			std::array<float, 4> e{0, 0, 0, 0};
 		};
 
 		vec4() = default;
-		vec4(double v);
+		vec4(float v);
 		vec4(const vec4& v) = default;
 		vec4(const glm::vec4 v);
 
-		vec4(double e0, double e1, double e2, double e3) : e{ e0, e1, e2, e3 } {}
-		vec4(std::initializer_list<double> args);
+		vec4(float e0, float e1, float e2, float e3) : e{ e0, e1, e2, e3 } {}
+		vec4(std::initializer_list<float> args);
 
-		vec4(const std::array<double, 4>&v) : e(v) {};
+		vec4(const std::array<float, 4>&v) : e(v) {};
 		vec4(const vec3& v3);
-		vec4(const vec3& v3, double w);
+		vec4(const vec3& v3, float w);
 
 		vec4& operator=(const vec4& v) = default;
 
-		inline double operator [] (int i) const { return e[i]; }
-		inline double& operator [] (int i) { return e[i]; }
+		inline float operator [] (int i) const { return e[i]; }
+		inline float& operator [] (int i) { return e[i]; }
 		
 		bool operator == (const vec4& other);
 		vec4 operator + (const vec4& v) const;
 		vec4 operator - () const;
 		vec4& operator += (const vec4& v);
-		vec4& operator *= (const double& f);
-		vec4& operator /= (const double t);
+		vec4& operator *= (const float& f);
+		vec4& operator /= (const float t);
 
 		friend inline vec4 operator - (vec4 u, vec4 v);
 		friend inline vec4 operator * (const vec4& u, const vec4& v);
-		friend inline vec4 operator * (double t, const vec4& v);
+		friend inline vec4 operator * (float t, const vec4& v);
 
-		inline std::array<double, 4> get_array() const { return e; }
-		double lenght() const;
-		double lenght_squared() const;
+		inline std::array<float, 4> get_array() const { return e; }
+		float lenght() const;
+		float lenght_squared() const;
 
-		double dot(const vec4& v) const;
+		float dot(const vec4& v) const;
 		vec4 unit_vector() const;
 		friend inline std::ostream& operator << (std::ostream& out, const vec4& v);
 	};
